fix: separate web interface instance
