+++ bot.py
@@ -5,7 +5,9 @@ from configparser import ConfigParser
 import time
 import schedule
 from jinja2 import Template, Environment, FileSystemLoader
-from web_interface import WebInterface
+from interface.pcmax import Pcmax
+from interface.wikipedia import Wikipedia
+from interface.livedoor import Livedoor
 from weekdays import WEEKDAYS
 
 PURE = '1'
@@ -25,26 +27,27 @@ def generate_from_template(directory, template_file, data):
     template = env.get_template(template_file)
     return str(template.render(data))
 
-def generate_data_wrapper(web, pattern):
+def generate_data_wrapper(pattern):
     # pattern毎に必要なデータを取得する
     if pattern == 'morning':
-        return generate_data_for_morning(web)
+        return generate_data_for_morning()
     elif pattern == 'night':
-        return generate_data_for_night(web)
+        return generate_data_for_night()
     else:
         # patternに一致しない場合は例外で終了
         raise ValueError('Pattern "%s" is not defined')
 
-def generate_data_for_morning(web):
+def generate_data_for_morning():
     # 朝のつぶやき用の情報を取得する
     today = datetime.date.today()
     month = str(today.month)
     day = str(today.day)
     weekday = WEEKDAYS[today.weekday()]
     # WikipediaのAPIから「今日は何の日」かを取得する
+    wikipedia = Wikipedia()
     url = config.get('event', 'url').replace('MONTH', month)
     date = '%s月%s日' % (month, day)
-    event = web.get_today_event(url, date)
+    event = wikipedia.get_today_event(url, date)
     if event_pattern.search(event):
         # 歴史上の出来事なのか記念日なのかで語尾を使い分ける
         end_word = 'ガアッタ日ダソウデス。'
@@ -57,7 +60,7 @@ def generate_data_for_morning(web):
             'event': event,
             'end_word': end_word}
 
-def generate_data_for_night(web):
+def generate_data_for_night():
     # 夜のつぶやき用の情報を取得する
     today = datetime.date.today()
     tommorrow = today + datetime.timedelta(days=1)
@@ -65,15 +68,16 @@ def generate_data_for_night(web):
     day = str(tommorrow.day)
     weekday = WEEKDAYS[tommorrow.weekday()]
     # LivedoorのAPIから東京/大阪/名古屋の天気と気温を取得する
+    livedoor = Livedoor()
     url = config.get('weather', 'url')
     tokyo = config.get('weather', 'tokyo')
     osaka = config.get('weather', 'osaka')
     nagoya = config.get('weather', 'nagoya')
     sapporo = config.get('weather', 'sapporo')
-    forecast_tokyo = web.get_today_weather(url, tokyo)
-    forecast_osaka = web.get_today_weather(url, osaka)
-    forecast_nagoya = web.get_today_weather(url, nagoya)
-    forecast_sapporo = web.get_today_weather(url, sapporo)
+    forecast_tokyo = livedoor.get_today_weather(url, tokyo)
+    forecast_osaka = livedoor.get_today_weather(url, osaka)
+    forecast_nagoya = livedoor.get_today_weather(url, nagoya)
+    forecast_sapporo = livedoor.get_today_weather(url, sapporo)
     return {'pattern': 'night',
             'month': month,
             'day': day,
@@ -92,19 +96,21 @@ def generate_data_for_night(web):
             'min_sapporo': forecast_sapporo['min']}
 
 def tweet_wrapper(pattern):
-    # Web用のインスタンスを生成してつぶやきをポストする
+    # PCMAXにログインしてつぶやきを投稿するためのインスタンスを生成し、つぶやきをポストする
     print('/*--- tweet begin ---*/')
     login_url = config.get('pcmax', 'login_url')
     login_user = os.environ.get('LOGIN_USER')
     login_password = os.environ.get('LOGIN_PASSWORD')
-    web = WebInterface(login_url, login_user, login_password)
+
+    pcmax = Pcmax(login_url, login_user, login_password)
     tweet_url = config.get('pcmax', 'tweet_url').replace('ROOM', PURE)
-    tweet_input = web.get_tweet_input(tweet_url)
-    data = generate_data_wrapper(web, pattern)
+    tweet_input = pcmax.get_tweet_input(tweet_url)
+
+    data = generate_data_wrapper(pattern)
     tweet_text = generate_from_template('template', '%s.j2'%pattern, data).encode('shift_jis')
     print(datetime.datetime.now().strftime('%Y/%m/%d %H:%M'))
     print(tweet_text.decode('shift_jis'))
-    print(web.post_tweet(tweet_url, tweet_input, tweet_text).status_code)
+    print(pcmax.post_tweet(tweet_url, tweet_input, tweet_text).status_code)
     print('/*--- tweet end ---*/')
     print('')
     print('Bot is aliving...')
