+++ check_output.py
@@ -1,6 +1,5 @@
 import os
 from configparser import ConfigParser
-from web_interface import WebInterface
 from bot import generate_from_template, generate_data_wrapper, generate_data_for_morning, generate_data_for_night
 
 def main():
@@ -10,9 +9,8 @@ def main():
     login_url = config.get('pcmax', 'login_url')
     login_user = os.environ.get('LOGIN_USER')
     login_password = os.environ.get('LOGIN_PASSWORD')
-    web = WebInterface(login_url, login_user, login_password)
-    morning_data = generate_data_wrapper(web, 'morning')
-    night_data = generate_data_wrapper(web, 'night')
+    morning_data = generate_data_wrapper('morning')
+    night_data = generate_data_wrapper('night')
     moring_tweet_text = generate_from_template('template', 'morning.j2', morning_data).encode('shift_jis')
     night_tweet_text = generate_from_template('template', 'night.j2', night_data).encode('shift_jis')
